research/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pandas
numpy
matplotlib
kagglehub
scikit-learn
xgboost
seaborn
ipykernel
jupyter
